third_party/blink/web_tests/external/wpt/offscreen-canvas/text/2d.text.draw.space.collapse.nonspace.worker-expected.txt

This is a testharness.js-based test.
FAIL Non-space characters are not converted to U+0020 and collapsed assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
